# futura-tickets-admin-api/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/accounts/admin": {
      "get": {
        "description": "Retrieve all admin accounts for the authenticated promoter.",
        "operationId": "AccountController_getAdminAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of admin accounts retrieved successfully."
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get admin accounts",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/access/{event}": {
      "get": {
        "description": "Retrieve all access control accounts assigned to a specific event.",
        "operationId": "AccountController_getEventAccessAccounts",
        "parameters": [
          {
            "name": "event",
            "required": true,
            "in": "path",
            "description": "Event ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of access accounts retrieved successfully."
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get access accounts for event",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/admin/{account}": {
      "delete": {
        "description": "Delete an admin account by ID. Requires promoter authentication.",
        "operationId": "AccountController_deleteAdminAccounts",
        "parameters": [
          {
            "name": "account",
            "required": true,
            "in": "path",
            "description": "Account ID to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin account deleted successfully."
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token."
          },
          "404": {
            "description": "Account not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete admin account",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/create": {
      "post": {
        "description": "Create a new user account. Public endpoint for user registration.",
        "operationId": "AccountController_createAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "createAccount": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "Juan"
                      },
                      "lastName": {
                        "type": "string",
                        "example": "Pérez"
                      },
                      "email": {
                        "type": "string",
                        "example": "juan.perez@example.com"
                      },
                      "password": {
                        "type": "string",
                        "example": "SecurePass123!"
                      },
                      "phone": {
                        "type": "string",
                        "example": "+34600123456"
                      }
                    },
                    "required": [
                      "name",
                      "lastName",
                      "email",
                      "password"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully."
          },
          "400": {
            "description": "Invalid input data."
          },
          "409": {
            "description": "Email already exists."
          }
        },
        "summary": "Create new account",
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/accounts/login": {
      "post": {
        "description": "Authenticate a user with email and password. Returns user data with JWT token.",
        "operationId": "AccountController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns account data with JWT token."
          },
          "401": {
            "description": "Invalid credentials."
          }
        },
        "summary": "User login",
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/accounts/access/login": {
      "post": {
        "operationId": "AccountController_accessLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/accounts/address/{address}": {
      "get": {
        "operationId": "AccountController_getAccountByAddress",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/promoters/clients/{client}": {
      "get": {
        "operationId": "PromoterController_getPromoterUser",
        "parameters": [
          {
            "name": "client",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Promoter"
        ]
      }
    },
...
    "/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_getPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_deletePayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/payment-methods/{id}": {
      "get": {
        "operationId": "PaymentMethodsController_getPaymentMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PaymentMethods"
        ]
      },
      "delete": {
        "operationId": "PaymentMethodsController_deletePaymentMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PaymentMethods"
        ]
      }
    },
...
  },
  "info": {
    "title": "FuturaTickets Admin API",
    "description": "\n## FuturaTickets Admin API Documentation\n\nThis API provides comprehensive endpoints for managing events, tickets, orders, and users in the FuturaTickets platform.\n\n### Getting Started\n\n1. **Authentication**: Most endpoints require a JWT Bearer token. Obtain a token by logging in via `/auth/login`.\n2. **Authorization**: Click the \"Authorize\" button above and enter your Bearer token.\n3. **Roles**: Different endpoints require different roles (ADMIN, PROMOTER, USER).\n\n### API Features\n\n- Event Management (CRUD operations)\n- Order Processing & Payment Integration\n- Ticket Sales & Validation\n- User & Account Management\n- Promoter Management\n- Email Notifications\n- WebSocket Support for Real-time Updates\n\n### Rate Limits\n\n| Endpoint Type | Requests per Minute |\n|--------------|---------------------|\n| Authentication | 10 |\n| Read Operations | 100 |\n| Write Operations | 50 |\n\n### Error Codes\n\n| Code | Description |\n|------|-------------|\n| 400 | Bad Request - Invalid input |\n| 401 | Unauthorized - Missing or invalid token |\n| 403 | Forbidden - Insufficient permissions |\n| 404 | Not Found - Resource doesn't exist |\n| 429 | Too Many Requests - Rate limit exceeded |\n| 500 | Internal Server Error |\n  ",
    "version": "1.0.0",
    "contact": {
      "name": "FuturaTickets Support",
      "url": "https://futuratickets.com/support",
...
    },
...
    },
...
  },
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter your JWT token obtained from the login endpoint",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@futuratickets.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "SecurePassword123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  },
...
}